# prevent expensive compilation of map / splatting