@timholy 2019-03-28
Absolutely, but please check my belated comment https://github.com/JuliaDebug/JuliaInterpreter.jl/commit/c1b99a1ea123d0c528aa32e62aff334c7ab709e5#r32949892 first.