@KristofferC 2019-03-28
Ok to tag or do you want to try out master for a while longer with all the compiled ccall stuff?